move base command class
